adds cast refs #61
svn merge -r 1882:1883 http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui/WORK_01

git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@1892 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ common/textfile.c
@@ -565,7 +565,7 @@ const void		*buf;
 		flushfile(tf);
 		tf->mode = TFMODE_WRITE;
 	}
-	leng = OEMSTRLEN(buffer);
+	leng = (UINT)OEMSTRLEN(buffer);
 #if defined(SUPPORT_TEXTCNV)
 	if (tf->fromoem != NULL) {
 		leng = (tf->fromoem)(tf->cnvbuf, tf->cnvbufsize / tf->width,
